feat: Add Patient Medications feature
- Created migration to add PatientMedications table with necessary fields and relationships.
- Implemented MedicationInputModel for capturing medication details.
- Added PatientMedication model to represent medication records.
- Developed IPatientRiskEvaluator interface and PatientRiskEvaluator service for evaluating patient risk based on vital signs, alerts, and welfare checks.
- Created DashboardStatsViewModel for dashboard statistics.
- Designed Medications view for facility nurses to log patient medications.
- Introduced _RiskBadge partial view to display patient risk levels.

+++ Controllers/FacilityNurseController.cs
@@ -149,7 +149,7 @@ public class FacilityNurseController : Controller
 
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> DiscontinueMedication(int medicationId, int patientId, CancellationToken cancellationToken)
+    public async Task<IActionResult> DiscontinueMedication(long medicationId, int patientId, CancellationToken cancellationToken)
     {
         var medication = await _dbContext.PatientMedications
             .SingleOrDefaultAsync(
